+++ include/cinolib/drawable_isocontour.cpp
@@ -70,7 +70,7 @@ template<class M, class V, class E, class P>
 CINO_INLINE
 void DrawableIsocontour<M,V,E,P>::draw(const float scene_size) const
 {
-    double cylind_rad = scene_size * 0.002 * thickness;
+    float cylind_rad = float(scene_size * 0.002 * thickness);
 
     for(vec3d p : this->segs) draw_sphere(p, cylind_rad, color);
